refactor(amazonqFeatureDev): add i18n helper in shared folder
