+++ packages/core/src/shared/i18n-helper.ts
@@ -3,8 +3,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import translations from '../../../package.nls.json'
-import { localize } from '../../shared/utilities/vsCodeUtils'
+import translations from '../../package.nls.json'
+import { localize } from './utilities/vsCodeUtils'
 
 export const i18n = (code: string) => {
     const translationsTyped = translations as { [key: string]: string }
